+++ packages/clerk-js/src/ui/localization/localizationModifiers.ts
@@ -19,11 +19,17 @@ const weekday = (val: Date | string | number, locale?: string, weekday?: 'long' 
   }
 };
 
+/**
+ * Returns a short date string if the current year is the same as the date's year,
+ * otherwise returns a long date string.
+ */
 const shortDate = (val: Date | string | number, locale?: string) => {
+  const date = normalizeDate(val);
   try {
     return new Intl.DateTimeFormat(locale || 'en-US', {
       month: 'short',
       day: 'numeric',
+      ...(date.getFullYear() !== new Date().getFullYear() ? { year: 'numeric' } : {}),
     }).format(normalizeDate(val));
   } catch (e) {
     console.warn(e);