+++ packages/localizations/src/es-ES.ts
@@ -20,6 +20,10 @@ export const esES: LocalizationResource = {
     checkout: {
       description__paymentSuccessful: undefined,
       description__subscriptionSuccessful: undefined,
+      emailForm: {
+        subtitle: undefined,
+        title: undefined,
+      },
       lineItems: {
         title__invoiceId: undefined,
         title__paymentMethod: undefined,
@@ -616,7 +620,8 @@ export const esES: LocalizationResource = {
     form_username_invalid_character: 'El nombre de usuario contiene caracteres inválidos.',
     form_username_invalid_length: 'El nombre de usuario debe tener entre 3 y 20 caracteres.',
     identification_deletion_failed: 'No puedes eliminar tu última identificación.',
-    not_allowed_access: 'Acceso no permitido.',
+    not_allowed_access:
+      "La dirección de correo electrónico o el número de teléfono no está permitido para registrarse. Esto puede deberse al uso de '+', '=', '#' o '.' en tu dirección de correo electrónico, el uso de un dominio conectado a un servicio de correo electrónico temporal o la exclusión explícita. Si cree que se trata de un error, póngase en contacto con el soporte.",
     organization_domain_blocked: 'Este es un dominio bloqueado, por favor usa otro.',
     organization_domain_common: 'Este es un dominio común, por favor usa otro.',
     organization_domain_exists_for_enterprise_connection: undefined,
@@ -638,7 +643,8 @@ export const esES: LocalizationResource = {
       requireUppercase: 'La contraseña debe contener al menos una letra mayúscula.',
       sentencePrefix: 'Tu contraseña debe cumplir con los requisitos de complejidad.',
     },
-    phone_number_exists: undefined,
+    phone_number_exists: 'Este número de teléfono ya está en uso. Por favor, inténtelo con otro.',
+    session_exists: 'Ya has iniciado sesión',
     web3_missing_identifier: undefined,
     zxcvbn: {
       couldBeStronger: undefined,